# docs/api/admin-api/openapi-full.yaml
...
openapi: 3.0.3
info:
  version: v0.2.3
  title: Private API print-shop
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Andrey Nazarov
  description: |-
    Private REST API проекта print-shop

    ### Пользовательские ограничения полей и ошибки

    #### Ограничения полей
    - required - поле обязательно для заполнения;
    - omitempty - поле может быть не указано (не будет использоваться методом, в который было передано);
    - unique - уникальное значение поля;
    - min=N - поле должно быть не менее N символов;
    - max=N - поле должно быть не более N символов;
    - gte=N - числовое поле должно быть равно или более N;
    - lte=N - числовое поле должно быть равно или менее N;
    - enum - поле должно содержать одно из ENUM значений;
    - UUID - поле формата UUID;
    - pattern=P - поле должно соответствовать регулярному выражению P;

    #### Ошибки
    - ErrVersionInvalid - если передаваемая версия объекта не совпала с текущей версией объекта.\
      Как правило, это означает, что объект был ранее изменён другим процессом;
    - ErrSwitchStatusRejected - перевод в указанный статус объекта отклонён.\
      WorkFlow объекта запрещает переключение в указанный статус;
servers:
  - url: http://localhost:8080/adm
    description: Local Host
...
paths:
...
  /v1/dictionaries/material-types:
    get:
      summary: Метод получения списка типов материала
      description: |-
        Метод предназначен для получения списка всех типов материала, сохраненных в БД

        Фильтр `App.Request.Query.Filter.SearchText` использует для поиска следующие поля:
        - caption;

        Возможные значения `App.Request.Query.ListSorter.SortField`:
        - caption,default,asc;
        - createdAt;
        - updatedAt;

        Возможные значения `App.Request.Query.ListPager.PageSize`:
        - от 1 до 1000 (default: 50)

      operationId: getMaterialTypeList
      tags:
        - Dictionaries.MaterialType

      security:
        - bearerAuth: []

      parameters:
        - $ref: "#/components/parameters/App.Request.Header.AcceptLanguage"
        - $ref: "#/components/parameters/App.Request.Header.CorrelationID"
        - $ref: "#/components/parameters/App.Request.Query.Filter.SearchText"
        - $ref: "#/components/parameters/App.Request.Query.Filter.Statuses"
        - $ref: "#/components/parameters/App.Request.Query.ListSorter.SortField"
        - $ref: "#/components/parameters/App.Request.Query.ListSorter.SortDirection"
        - $ref: "#/components/parameters/App.Request.Query.ListPager.PageIndex"
        - $ref: "#/components/parameters/App.Request.Query.ListPager.PageSize"

      responses:
        "200":
          description: Получен список типов материала
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: "#/components/schemas/Dictionaries.MaterialTypes.AdminAPI.Response.Model.MaterialTypeList"
        "401":
          $ref: "#/components/responses/App.ResponseJson.Error401"
        "403":
          $ref: "#/components/responses/App.ResponseJson.Error403"
        "500":
          $ref: "#/components/responses/App.ResponseJson.Error5XX"

    post:
      summary: Метод добавления нового типа материала
      description: |-
        Метод предназначен для создания информации о новом типе материала в БД

        **Ограничения и ошибки полей:**
        - caption: required,max=64

      operationId: createMaterialType
      tags:
        - Dictionaries.MaterialType

      security:
        - bearerAuth: []

      parameters:
        - $ref: "#/components/parameters/App.Request.Header.AcceptLanguage"
        - $ref: "#/components/parameters/App.Request.Header.CorrelationID"

      requestBody:
        required: true
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/Dictionaries.MaterialTypes.AdminAPI.Request.Model.CreateMaterialType"

      responses:
        "201":
          description: Информация о типе материала успешно добавлена
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: "#/components/schemas/App.Response.Model.SuccessCreatedItemInt32"
        "400":
          $ref: "#/components/responses/App.ResponseJson.Error400"
        "401":
          $ref: "#/components/responses/App.ResponseJson.Error401"
        "403":
          $ref: "#/components/responses/App.ResponseJson.Error403"
        "500":
          $ref: "#/components/responses/App.ResponseJson.Error5XX"
  /v1/dictionaries/material-types/{typeId}:
    get:
      summary: Метод получения типа материала
      description: Метод предназначен для получения информации о типе материала из БД
      operationId: getMaterialType
      tags:
        - Dictionaries.MaterialType

      security:
        - bearerAuth: []

      parameters:
        - $ref: "#/components/parameters/App.Request.Header.AcceptLanguage"
        - $ref: "#/components/parameters/App.Request.Header.CorrelationID"
        - $ref: "#/components/parameters/Dictionaries.MaterialTypes.AdminAPI.Request.Path.TypeID"

      responses:
        "200":
          description: Получена информация о типе материала
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: "#/components/schemas/Dictionaries.MaterialTypes.AdminAPI.Response.Model.MaterialType"
        "401":
          $ref: "#/components/responses/App.ResponseJson.Error401"
        "403":
          $ref: "#/components/responses/App.ResponseJson.Error403"
        "404":
          $ref: "#/components/responses/App.ResponseJson.Error404"
        "500":
          $ref: "#/components/responses/App.ResponseJson.Error5XX"

    put:
      summary: Метод сохранения типа материала
      description: |-
        Метод предназначен для сохранения информации о типе материала в БД

        **Ограничения и ошибки полей:**
        - caption: required,max=64

      operationId: storeMaterialType
      tags:
        - Dictionaries.MaterialType

      security:
        - bearerAuth: []

      parameters:
        - $ref: "#/components/parameters/App.Request.Header.AcceptLanguage"
        - $ref: "#/components/parameters/App.Request.Header.CorrelationID"
        - $ref: "#/components/parameters/Dictionaries.MaterialTypes.AdminAPI.Request.Path.TypeID"

      requestBody:
        required: true
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/Dictionaries.MaterialTypes.AdminAPI.Request.Model.StoreMaterialType"

      responses:
        "204":
          description: Информация о типе материала успешно сохранена
        "400":
          $ref: "#/components/responses/App.ResponseJson.Error400"
        "401":
          $ref: "#/components/responses/App.ResponseJson.Error401"
        "403":
          $ref: "#/components/responses/App.ResponseJson.Error403"
        "404":
          $ref: "#/components/responses/App.ResponseJson.Error404"
        "500":
          $ref: "#/components/responses/App.ResponseJson.Error5XX"

    delete:
      summary: Метод удаления типа материала
      description: Метод предназначен для удаления информации о типе материала из БД
      operationId: removeMaterialType
      tags:
        - Dictionaries.MaterialType

      security:
        - bearerAuth: []

      parameters:
        - $ref: "#/components/parameters/App.Request.Header.AcceptLanguage"
        - $ref: "#/components/parameters/App.Request.Header.CorrelationID"
        - $ref: "#/components/parameters/Dictionaries.MaterialTypes.AdminAPI.Request.Path.TypeID"

      responses:
        "204":
          description: Информация о типе материала успешно удалена
        "401":
          $ref: "#/components/responses/App.ResponseJson.Error401"
        "403":
          $ref: "#/components/responses/App.ResponseJson.Error403"
        "404":
          $ref: "#/components/responses/App.ResponseJson.Error404"
        "500":
          $ref: "#/components/responses/App.ResponseJson.Error5XX"
  /v1/dictionaries/material-types/{typeId}/status:
    patch:
      summary: Метод изменения статуса типа материала
      description: |-
        Метод предназначен для изменения статуса типа материала в БД

        **Ограничения и ошибки полей:**
        - tagVersion: required,gte=1 | ErrVersionInvalid
        - status: required,enum | ErrSwitchStatusRejected

      operationId: changeMaterialTypeStatus
      tags:
        - Dictionaries.MaterialType

      security:
        - bearerAuth: []

      parameters:
        - $ref: "#/components/parameters/App.Request.Header.AcceptLanguage"
        - $ref: "#/components/parameters/App.Request.Header.CorrelationID"
        - $ref: "#/components/parameters/Dictionaries.MaterialTypes.AdminAPI.Request.Path.TypeID"

      requestBody:
        required: true
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/App.Request.Model.ChangeStatus"

      responses:
        "204":
          description: Статус типа материала успешно изменён
        "400":
          $ref: "#/components/responses/App.ResponseJson.Error400"
        "401":
          $ref: "#/components/responses/App.ResponseJson.Error401"
        "403":
          $ref: "#/components/responses/App.ResponseJson.Error403"
        "404":
          $ref: "#/components/responses/App.ResponseJson.Error404"
        "500":
          $ref: "#/components/responses/App.ResponseJson.Error5XX"
...
components:
  parameters:
    App.Request.Header.AcceptLanguage:
      description: https://developer.mozilla.org/ru/docs/Web/HTTP/Headers/Accept-Language
      name: Accept-Language
      in: header
      required: false
      schema:
        type: string
        default: "ru-RU"
      example: "ru;q=0.9, fr-CH, fr;q=0.8, en;q=0.7, *;q=0.5"
    App.Request.Header.CorrelationID:
      description: Уникальный идентификатор клиента
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      example: "1f0b7925-cee6-4e44-8ea3-4e585d12212d"
    App.Request.Query.Filter.SearchText:
      description: Поисковая строка
      name: query
      in: query
      required: false
      schema:
        type: string
        maxLength: 128
      example: "перфокарта"
    App.Request.Query.Filter.Statuses:
      description: Список статусов через "," (см. App.Enum.Status)
      name: statuses
      in: query
      required: false
      schema:
        type: string
        default: "ENABLED"
      example: "DRAFT,ENABLED"
    App.Request.Query.ListPager.PageIndex:
      description: Индекс страницы списка с записями
      name: pageIndex
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        maximum: 1024
        default: 0
      example: 2
    App.Request.Query.ListPager.PageSize:
      description: Количество записей на страницу списка
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1024
        default: 50
      example: 70
    App.Request.Query.ListSorter.SortField:
      description: Название колонки, по которой происходит сортировка
      name: sortField
      in: query
      required: false
      schema:
        type: string
      example: "caption"
    App.Request.Query.ListSorter.SortDirection:
      description: Направление сортировки
      name: sortDirection
      in: query
      required: false
      schema:
        description: |-
          Направление сортировки:
          - ASC - по возрастанию;
          - DESC - по убыванию;
        type: string
        enum:
          - ASC
          - DESC
        default: "ASC"
      example: "DESC"
...
    Dictionaries.MaterialTypes.AdminAPI.Request.Path.TypeID:
      description: Идентификатор типа материала
      name: typeId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      example: 1000
...
  schemas:
    App.Enum.Status:
      description: |-
        Текущий статус записи:
        - DRAFT - черновик;
        - ENABLED - используется;
        - DISABLED - отключена;
      type: string
      enum:
        - DRAFT
        - ENABLED
        - DISABLED
      example: "ENABLED"
...
    App.Field.Caption:
      description: Название записи
      type: string
      maxLength: 128
      example: "Название"
    App.Field.DateTimeCreatedAt:
      description: Дата и время создания записи (UTC)
      type: string
      format: date-time
      example: "2023-01-01T12:00:00.000+02:00"
    App.Field.DateTimeUpdatedAt:
      description: Дата и время обновления записи (UTC)
      type: string
      format: date-time
      example: "2023-01-01T12:00:00.000+02:00"
...
    App.Field.ListPager.Total:
      description: Количество записей в БД с учётом указанных в запросе фильтров
      type: integer
      minimum: 0
      example: 584
...
    App.Field.TagVersion:
      description: Текущая версия записи
      type: integer
      minimum: 1
      example: 5
...
    App.Request.Model.ChangeStatus:
      description: |-
        DTO изменения статуса записи

        Возможные варианты переключений:
        - DRAFT -> ENABLED, DISABLED;
        - ENABLED -> DISABLED;
        - DISABLED -> ENABLED;
      type: object
      required:
        - tagVersion
        - status
      properties:
        tagVersion:
          $ref: "#/components/schemas/App.Field.TagVersion"
        status:
          $ref: "#/components/schemas/App.Enum.Status"
...
    App.Response.Model.ErrorList:
      description: DTO списка пользовательских ошибок 400
      type: array
      items:
        $ref: "#/components/schemas/App.Response.Model.ErrorAttribute"
    App.Response.Model.ErrorAttribute:
      description: DTO пользовательской ошибки 400
      type: object
      required:
        - id
        - value
      properties:
        id:
          description: Название атрибута, где произошла ошибка
          type: string
          example: "attributeName"
        value:
          description: Описание ошибки
          type: string
          example: "Атрибут не может быть пустым | Указано недопустимое значение атрибута"
    App.Response.Model.ErrorDetails: # RFC 7807 (Problem Details for HTTP APIs)
      description: DTO следующих ошибок - 401, 403, 404, 5XX
      type: object
      required:
        - title
        - details
        - request
        - time
      properties:
        title:
          description: Краткое описание проблемы, понятное человеку
          type: string
          example: "Resource not found"
        details:
          description: Описание конкретно возникшей ошибки, понятное человеку
          type: string
          example: "Resource [User] with id = [123456] not found. You MUST use PUT to add a resource instead of GET"
        request:
          description: Метод и URL запроса
          type: string
          example: "PUT /users/123456"
        time:
          description: Время возникновения ошибки с точностью до миллисекунд (UTC)
          type: string
          format: date-time
          example: "2023-01-01T12:00:00.000+02:00"
        errorTraceId:
          description: Идентификатор конкретного возникновения ошибки
          type: string
          example: "5add1be1-90ab5d42-02fa8b1f-672503f2"
...
    App.Response.Model.SuccessCreatedItemInt32:
      description: DTO результата создания записи (int32 ID)
      type: object
      required:
        - id
      properties:
        id:
          description: Идентификатор созданной записи
          type: integer
          format: int32
          example: 2147483647
...
    Custom.Field.Dictionaries.MaterialTypeID:
      description: Идентификатор типа материала (модуль - Dictionaries.MaterialType)
      type: integer
      format: int32
      minimum: 1
      example: 2
...
    Dictionaries.MaterialTypes.AdminAPI.Request.Model.CreateMaterialType:
      description: DTO добавления нового типа материала
      type: object
      required:
        - caption
      properties:
        caption:
          allOf:
            - $ref: "#/components/schemas/App.Field.Caption"
            - maxLength: 64
            - example: "Глянцевый"
    Dictionaries.MaterialTypes.AdminAPI.Request.Model.StoreMaterialType:
      description: DTO сохранения типа материала
      type: object
      required:
        - tagVersion
        - caption
      properties:
        tagVersion:
          $ref: "#/components/schemas/App.Field.TagVersion"
        caption:
          allOf:
            - $ref: "#/components/schemas/App.Field.Caption"
            - maxLength: 64
            - example: "Глянцевый"
    Dictionaries.MaterialTypes.AdminAPI.Response.Model.MaterialTypeList:
      description: DTO - список типов материала
      type: object
      required:
        - items
        - total
      properties:
        items:
          description: Список записей ограниченный фильтром и pageSize
          type: array
          items:
            $ref: "#/components/schemas/Dictionaries.MaterialTypes.AdminAPI.Response.Model.MaterialType"
        total:
          $ref: "#/components/schemas/App.Field.ListPager.Total"
    Dictionaries.MaterialTypes.AdminAPI.Response.Model.MaterialType:
      description: DTO - тип материала
      type: object
      required:
        - id
        - tagVersion
        - caption
        - status
        - createdAt
      properties:
        id: # typeId
          $ref: "#/components/schemas/Custom.Field.Dictionaries.MaterialTypeID"
        tagVersion:
          $ref: "#/components/schemas/App.Field.TagVersion"
        caption:
          allOf:
            - $ref: "#/components/schemas/App.Field.Caption"
            - maxLength: 64
            - example: "Глянцевый"
        status:
          $ref: "#/components/schemas/App.Enum.Status"
        createdAt:
          $ref: "#/components/schemas/App.Field.DateTimeCreatedAt"
        updatedAt:
          $ref: "#/components/schemas/App.Field.DateTimeUpdatedAt"
...
  responses:
    App.ResponseJson.Error400:
      description: Некорректные входные данные. Возвращается список атрибутов с ошибками
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: "#/components/schemas/App.Response.Model.ErrorList"
    App.ResponseJson.Error404:
      description: Запрашиваемый ресурс не был найден
      content:
        application/problem+json;charset=UTF-8:
          schema:
            $ref: "#/components/schemas/App.Response.Model.ErrorDetails"
    App.ResponseJson.Error5XX:
      description: Внутренняя ошибка сервера
      content:
        application/problem+json;charset=UTF-8:
          schema:
            $ref: "#/components/schemas/App.Response.Model.ErrorDetails"
    App.ResponseJson.Error401:
      description: Пользователь не был аутентифицирован
      content:
        application/problem+json;charset=UTF-8:
          schema:
            $ref: "#/components/schemas/App.Response.Model.ErrorDetails"
    App.ResponseJson.Error403:
      description: Пользователю с текущими правами доступ отклонён
      content:
        application/problem+json;charset=UTF-8:
          schema:
            $ref: "#/components/schemas/App.Response.Model.ErrorDetails"
  securitySchemes:
    bearerAuth:
      description: https://www.rfc-editor.org/rfc/rfc6750
      type: http
      scheme: bearer
      bearerFormat: JWT